Preparing Information Page 2
- nothing

+++ lib/model/home_body_model/chart_home_model.dart
@@ -1,6 +1,6 @@
 class ChartDataModel {
   String coreName1;
-  int clockSpeed;
+  double clockSpeed;
 
   ChartDataModel({required this.coreName1, required this.clockSpeed});
 }
